requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
coverage==4.5.3
Django==2.2
factory-boy==2.11.1
Faker==1.0.5
more-itertools==7.0.0
Pillow==6.0.0
pluggy==0.9.0
py==1.8.0
pytest==4.4.1
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-pythonpath==0.7.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
